add hero shot
