+++ client/src/components/home/Home.css
@@ -1,0 +1,15 @@
+.hero {
+    height: calc(100vh - 180px);
+    background: rgb(6,110,226);
+    background: url("https://www.transparenttextures.com/patterns/back-pattern.png"), linear-gradient(170deg, rgba(6,110,226,1) 0%, rgba(30,184,214,1) 100%);
+}
+
+.hero-text {
+    color: rgb(252, 252, 252);
+    text-shadow: 1px 1px 3px rgba(25, 25, 25, 0.15);
+}
+
+.hero-text > h1 {
+    font-size: 4em;
+}
+
